requirement.txt
phidata
python-dotenv
sqlalchemy
pgvector
psycopg[binary]
pypdf
typer
google-generativeai
streamlit
duckduckgo-search